requirements.txt
pillow==11.3.0
PyAutoGUI==0.9.54

# Dependências indiretas do PyAutoGUI
MouseInfo==0.1.3
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0